// openmetadata-ui/src/main/resources/ui/src/components/common/EntityImport/CsvWorkflowHeader/CsvWorkflowHeader.interface.ts
/*
 *  Copyright 2026 Collate.
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *  http://www.apache.org/licenses/LICENSE-2.0
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */

import type { ReactNode } from 'react';
import type { TitleBreadcrumbProps } from '../../TitleBreadcrumb/TitleBreadcrumb.interface';

export interface CsvWorkflowStep {
  name: ReactNode;
  step: number;
}

export interface CsvWorkflowHeaderProps {
  activeStep: number;
  breadcrumbList: TitleBreadcrumbProps['titleLinks'];
  currentLabel: string;
  description: ReactNode;
  steps: CsvWorkflowStep[];
  title: ReactNode;
}